Move source files under src/
