+++ src/__tests__/explode.js
@@ -1,4 +1,4 @@
-const particl = require('..');
+const particl = require('../particl');
 
 describe('.explode()', () => {
   let p;
