+++ src/__tests__/has.js
@@ -1,4 +1,4 @@
-const particl = require('..');
+const particl = require('../particl');
 
 describe('.has()', () => {
   let p;
