+++ src/__tests__/need.js
@@ -1,4 +1,4 @@
-const particl = require('..');
+const particl = require('../particl');
 
 describe('.need()', () => {
   let p;
