+++ src/__tests__/next.js
@@ -1,4 +1,4 @@
-const particl = require('..');
+const particl = require('../particl');
 
 describe('.provide()', () => {
   let p;
