+++ src/__tests__/on.js
@@ -1,4 +1,4 @@
-const particl = require('..');
+const particl = require('../particl');
 
 describe('.on()', () => {
   let p;
