+++ src/mixins/__tests__/batchers.js
@@ -1,4 +1,4 @@
-const particl = require('../..');
+const particl = require('../../particl');
 const batchersMixin = require('../batchers');
 
 describe('mixins/batchers', () => {
